refactor: 롬복을 제거하라

+++ src/main/java/prgrms/project/shorturl/dto/ShortUrlRedirectResponse.java
@@ -2,6 +2,6 @@ package prgrms.project.shorturl.dto;
 
 public record ShortUrlRedirectResponse(
     String originUrl,
-    Long requestCount
+    int requestCount
 ) {
 }
